requirements.txt
flask==2.3.3
flask-login==0.6.3
flask-cors==4.0.0
flask-socketio==5.3.6
flask_sqlalchemy==3.1.1
PyMySQL==1.1.0
sqlalchemy==2.0.30
redis==5.0.3
dogpile.cache==1.1.8
kafka-python==2.0.2
cryptography==42.0.5
python-socketio==5.11.2
python-engineio==4.9.1
